Code refactored; Model added;
