+++ NoteApp/Model/Note.swift
@@ -1,7 +1,8 @@
 import UIKit
 import RealmSwift
 
-class Note:Object {
+class Note: Object {
+    
     @objc dynamic var title: String? = ""
     @objc dynamic var body: String? = ""
     @objc dynamic var isExpanded = false
